Move ENI/netplan generation from cloudconfig/containerinit to cloudconfig/cloudinit - it can be used by providers too, not only containers
