+++ cloudconfig/cloudinit/export_test.go
@@ -1,0 +1,11 @@
+// Copyright 2018 Canonical Ltd.
+// Copyright 2015 Cloudbase Solutions SRL
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package cloudinit
+
+var (
+	NetworkInterfacesFile       = &networkInterfacesFile
+	SystemNetworkInterfacesFile = &systemNetworkInterfacesFile
+	JujuNetplanFile             = &jujuNetplanFile
+)
