--- cloudconfig/containerinit/export_test.go
@@ -1,12 +1,0 @@
-// Copyright 2015 Canonical Ltd.
-// Copyright 2015 Cloudbase Solutions SRL
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package containerinit
-
-var (
-	NetworkInterfacesFile          = &networkInterfacesFile
-	SystemNetworkInterfacesFile    = &systemNetworkInterfacesFile
-	NewCloudInitConfigWithNetworks = newCloudInitConfigWithNetworks
-	JujuNetplanFile                = &jujuNetplanFile
-)
